+++ lib/lightning/workflows/job.ex
@@ -26,7 +26,6 @@ defmodule Lightning.Workflows.Job do
           __meta__: Ecto.Schema.Metadata.t(),
           id: Ecto.UUID.t() | nil,
           body: String.t() | nil,
-          enabled: boolean(),
           name: String.t() | nil,
           adaptor: String.t() | nil,
           credential: nil | Credential.t() | Ecto.Association.NotLoaded.t(),
@@ -38,7 +37,6 @@ defmodule Lightning.Workflows.Job do
   schema "jobs" do
     field :body, :string
 
-    field :enabled, :boolean, default: true
     field :name, :string
     field :adaptor, :string, default: "@openfn/language-common@latest"
 
@@ -66,7 +64,6 @@ defmodule Lightning.Workflows.Job do
         :inserted_at,
         :name,
         :body,
-        :enabled,
         :adaptor,
         :project_credential_id,
         :workflow_id
@@ -79,7 +76,7 @@ defmodule Lightning.Workflows.Job do
 
   def validate(changeset) do
     changeset
-    |> validate_required([:name, :body, :enabled, :adaptor])
+    |> validate_required([:name, :body, :adaptor])
     |> assoc_constraint(:workflow)
     |> validate_length(:name, max: 100)
     |> validate_format(:name, ~r/^[a-zA-Z0-9_\- ]*$/)